# user-management/src/main/java/com/bci/usermanagement/schema/response/UserResponse.java
package com.bci.usermanagement.schema.response;

import com.bci.usermanagement.entity.Telefono;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.UUID;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserResponse {
    private UUID id;
    private LocalDateTime creado;
    private LocalDateTime modificado;
    private LocalDateTime ultimoLogin;
    private String correo;
    private String nombre;
    private boolean activo;
}